Add function signatures for functions
That feature allows to find function corresponding to the given call.
For example, if we have the following function:

fun f(x, y, z) {
    return x + y - z
}

and interpreter sees the following call in the same scope:

f(2 * 3, 5)

then it should throw an error, because our language does not support any overloads, and the only function with the same name has 3 parameters (x, y, z), whereas only 2 arguments (2 * 3, 5) were passed to the call.

Moreover, with signatures we can produce more informative error messages.